+++ helm-cider.el
@@ -33,6 +33,7 @@
 (require 'cider)
 (require 'cl-lib)
 (require 'helm)
+(require 'helm-cider-cheatsheet)
 (require 'helm-cider-repl)
 (require 'helm-cider-spec)
 (require 'helm-cider-util)