+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## master (unreleased)
 
+### New features
+
+- [#10](https://github.com/clojure-emacs/helm-cider/pull/10): Port `clojure-cheatsheet`
+
 ## 0.4.0 (2018-01-20)
 
 ### New features